# node-api/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/get-bank-account": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/create-bank-account": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "personId": {
                  "type": "number",
                  "example": 1
                },
                "bankId": {
                  "type": "number",
                  "example": 1
                },
                "balance": {
                  "type": "number",
                  "example": 1
                },
                "currency": {
                  "type": "string",
                  "example": "EUR"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/transfer-funds": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "fromAccountId": {
                  "type": "number",
                  "example": 1
                },
                "toAccountId": {
                  "type": "number",
                  "example": 1
                },
                "amount": {
                  "type": "number",
                  "example": 1
                },
                "convert": {
                  "type": "boolean",
                  "example": false
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/withdraw-funds": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "accountId": {
                  "type": "number",
                  "example": 1
                },
                "amount": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/deposit-funds": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "accountId": {
                  "type": "number",
                  "example": 1
                },
                "currency": {
                  "type": "string",
                  "example": "EUR"
                },
                "amount": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/check-account-currencies": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "fromAccountId": {
                  "type": "number",
                  "example": 1
                },
                "recipientId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search-bank-accounts-by-person": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "personId",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}